+++ src/components/RecommendationsTable/RecommendationsTable.module.scss
